+++ src/main/java/racingcar/dao/game/GameDao.java
@@ -1,4 +1,4 @@
-package racingcar.dao;
+package racingcar.dao.game;
 
 public interface GameDao {
     int save(int tryCount);
