+++ src/main/java/racingcar/service/GameService.java
@@ -6,8 +6,8 @@ import java.util.Map;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import racingcar.controller.dto.GamePlayResponseDto;
-import racingcar.dao.CarDao;
-import racingcar.dao.GameDao;
+import racingcar.dao.car.CarDao;
+import racingcar.dao.game.GameDao;
 import racingcar.dao.dto.GameFinishedCarDto;
 import racingcar.model.Cars;
 import racingcar.model.TryCount;
