+++ src/test/java/racingcar/dao/InMemoryGameDaoTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import racingcar.dao.game.GameDao;
 
 @SpringBootTest
 class InMemoryGameDaoTest {
